code added
Jedis benchmarks for as many methods as can be thrown
in. To be tested with and without OpenCensus using JMH

However, they need a powerful computer to be run lest
take a long time. On my 97% SSD full and fully loaded
capacity Macbook, they take an hour at least to complete